requirements.txt
strands-agents>=1.0.0
strands-agents-tools>=0.2.0

# AWS Bedrock AgentCore (para ejecución con AgentCore CLI)
bedrock_agentcore

# AgentCore Memory integration

# Dependencias adicionales para el agente de servicio al cliente
python-dotenv>=0.19.0

# AWS Bedrock Knowledge Base (Requerido para agent4b_bedrock.py)
boto3>=1.34.94
botocore>=1.34.94

# Opcional: Para modelos en la nube
anthropic>=0.7.0
openai>=1.0.0
ollama>=0.1.0
strands-agents[openai]
